requirements.txt
# ── Backend ──────────────────────────────────────
fastapi
uvicorn[standard]
websockets
python-multipart

# ── Data & plotting ─────────────────────────────
pandas
numpy
plotly
folium
openpyxl

# ── Networking & utilities ──────────────────────
requests
jira
pyyaml
psutil
protobuf
tqdm
Pillow
pyserial

# ── System tray launcher ───────────────────────
pystray
